# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMDB API",
        "version": "3.0"
    },
    "servers": [
        {
            "url": "/api/v2",
            "description": "API v2"
        }
    ],
    "paths": {
...
        "/net/device": {
            "get": {
                "tags": [
                    "net/device"
                ],
                "summary": "NetDevice list",
                "operationId": "net/device",
                "responses": {
                    "200": {
                        "description": "NetDevice list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponceList"
                                        },
                                        {
                                            "properties": {
                                                "records": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/NetDevice"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTToken": []
                    }
                ]
            }
        },
        "/net/device/{id}": {
            "get": {
                "tags": [
                    "net/device"
                ],
                "summary": "NetDevice item",
                "operationId": "net/device/id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "NetDevice item",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponceItem"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/NetDevice"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTToken": []
                    }
                ]
            }
        },
        "/net/device/{id}/module": {
            "get": {
                "tags": [
                    "net/device"
                ],
                "summary": "NetDevice item modules",
                "operationId": "net/device/id/module",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "NetDevice item modules",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponceList"
                                        },
                                        {
                                            "properties": {
                                                "records": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/NetModule"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTToken": []
                    }
                ]
            }
        },
        "/net/dhcpsnoop": {
            "get": {
                "tags": [
                    "net/dhcpsnoop"
                ],
                "summary": "DHCP Snooping list",
                "operationId": "net/dhcpsnoop",
                "responses": {
                    "200": {
                        "description": "DHCP Snooping list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ResponceList"
                                        },
                                        {
                                            "properties": {
                                                "records": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Dhcp"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWTToken": []
                    }
                ]
            }
        },
        "/net/dhcpsnoop/create": {
            "post": {
                "tags": [
                    "net/dhcpsnoop"
                ],
                "summary": "DHCP Snooping Create",
                "operationId": "net/dhcpsnoop-create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "host_id": {
                                            "description": "Host ID",
                                            "type": "string"
                                        },
                                        "host_name": {
                                            "description": "Host Name",
                                            "type": "string"
                                        },
                                        "host_ip": {
                                            "description": "Host IP",
                                            "type": "string"
                                        },
                                        "host_status": {
                                            "description": "Host Status",
                                            "type": "string"
                                        },
                                        "host_snmp_status": {
                                            "description": "SNMP status",
                                            "type": "string"
                                        },
                                        "ip": {
                                            "description": "DHCP IP",
                                            "type": "string"
                                        },
                                        "mac": {
                                            "description": "DHCP MAC",
                                            "type": "string"
                                        },
                                        "vlan": {
                                            "description": "DHCP VLAN",
                                            "type": "string"
                                        },
                                        "port": {
                                            "description": "DHCP Port",
                                            "type": "string"
                                        },
                                        "ts": {
                                            "description": "UnixTimeStamp of last update",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Creating complete"
                    }
                },
                "security": [
                    {
                        "JWTToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ResponceItem": {
                "allOf": [
                    {
                        "properties": {
                            "status": {
                                "description": "Status code",
                                "type": "integer",
                                "default": 200
                            },
                            "message": {
                                "description": "Information message",
                                "type": "string"
                            },
                            "data": {
                                "description": "Data",
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            },
                            "extdata": {
                                "description": "ExtData",
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            },
                            "debug": {
                                "description": "Debug information",
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ResponceList": {
                "allOf": [
                    {
                        "properties": {
                            "status": {
                                "description": "Status code",
                                "type": "integer",
                                "default": 200
                            },
                            "message": {
                                "description": "Information message",
                                "type": "string"
                            },
                            "records": {
                                "description": "Records list",
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            },
                            "page": {
                                "description": "Actual page",
                                "type": "integer"
                            },
                            "limit": {
                                "description": "Limit on page",
                                "type": "integer"
                            },
                            "find_records": {
                                "description": "Find records",
                                "type": "integer"
                            },
                            "total_records": {
                                "description": "Total records in entity",
                                "type": "integer"
                            },
                            "debug": {
                                "description": "Debug information",
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Dhcp": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "host_id": {
                        "description": "Host ZABBIX ID",
                        "type": "string"
                    },
                    "host_name": {
                        "description": "Host name",
                        "type": "string"
                    },
                    "host_ip": {
                        "description": "Host IP",
                        "type": "string"
                    },
                    "host_status": {
                        "description": "Host status",
                        "type": "string"
                    },
                    "host_snmp_status": {
                        "description": "Host status by SNMP",
                        "type": "string"
                    },
                    "ip": {
                        "description": "DHCP IP",
                        "type": "string"
                    },
                    "mac": {
                        "description": "DHCP MAC",
                        "type": "string"
                    },
                    "vlan": {
                        "description": "DHCP VLAN",
                        "type": "string"
                    },
                    "port": {
                        "description": "DHCP Port",
                        "type": "string"
                    },
                    "ts": {
                        "description": "UnixTimeStamp of last update",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "NetDevice": {
                "properties": {
                    "id": {
                        "description": "NetDevice ID",
                        "type": "integer"
                    },
                    "upid": {
                        "description": "Parent NetDevice ID",
                        "type": "integer"
                    },
                    "type": {
                        "description": "NetDevice Type",
                        "type": "string"
                    },
                    "status": {
                        "description": "NetDevice status",
                        "type": "string"
                    },
                    "name": {
                        "description": "NetDevice hostname",
                        "type": "string"
                    },
                    "ip": {
                        "description": "NetDevice IP",
                        "type": "string"
                    },
                    "sn": {
                        "description": "NetDevice SN",
                        "type": "string"
                    },
                    "pn": {
                        "description": "NetDevice PN",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "NetModule": {
                "properties": {
                    "id": {
                        "description": "NetModule ID",
                        "type": "integer"
                    },
                    "type": {
                        "description": "NetModule Type",
                        "type": "string"
                    },
                    "status": {
                        "description": "NetModule status",
                        "type": "string"
                    },
                    "name": {
                        "description": "NetModule name",
                        "type": "string"
                    },
                    "ip": {
                        "description": "NetModule IP",
                        "type": "string"
                    },
                    "sn": {
                        "description": "NetModule SN",
                        "type": "string"
                    },
                    "pn": {
                        "description": "NetModule PN",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "JWTToken": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "JWTToken": []
        }
    ],
...
}